Add dataset_word_count.ipynb
Adds a new jupyter notebook with a prototypal algorithm for computing
the frequency of words appearing in the dataset names for each site.
The words are obtained by splitting the datatset name at every non
alphanumerical character and before each capital letter.

The last cell of the notebook requires the wordcloud package.