Add scripts to get data from McM public api
mcm_rest.py can be used to build an object able to call methods of
the McM publc api.
Running get_mcm_data.py is possible to query McM to get new info about
the datasets and write them on a new parquet file (this operation
is performed in chunks).